Remove unnecessary dependencies
