+++ package.json
@@ -35,7 +35,6 @@
     "@math.gl/web-mercator": "3.4.2",
     "@popperjs/core": "2.9.2",
     "@react-aria/button": "3.3.2",
-    "@react-aria/dialog": "3.1.3",
     "@react-aria/focus": "3.3.0",
     "@react-aria/i18n": "3.3.1",
     "@react-aria/interactions": "3.4.0",
@@ -44,7 +43,6 @@
     "@react-aria/slider": "3.0.7",
     "@react-aria/utils": "3.8.0",
     "@react-aria/visually-hidden": "3.2.2",
-    "@react-stately/overlays": "3.1.2",
     "@react-stately/searchfield": "3.1.2",
     "@react-stately/slider": "3.0.7",
     "@react-types/searchfield": "3.1.1",
@@ -53,7 +51,6 @@
     "@tailwindcss/aspect-ratio": "0.4.0",
     "@tailwindcss/forms": "0.5.1",
     "@tailwindcss/line-clamp": "0.4.0",
-    "@tippyjs/react": "4.2.5",
     "@vizzuality/layer-manager-layers-deckgl": "1.0.0-alpha.3",
     "@vizzuality/layer-manager-plugin-mapboxgl": "1.0.0-alpha.3",
     "@vizzuality/layer-manager-provider-carto": "1.0.0-alpha.3",
