+++ src/components/confirmation-prompt/component.tsx
@@ -18,7 +18,15 @@ export const ConfirmationPrompt: FC<ConfirmationPromptProps> = ({
   onAccept,
   onRefuse,
 }: ConfirmationPromptProps) => (
-  <Modal open={open} title={title} size="narrow" dismissable={dismissible} onDismiss={onDismiss}>
+  <Modal
+    open={open}
+    title={title}
+    size="default"
+    dismissable={dismissible}
+    onOpenChange={(o) => {
+      onDismiss(o);
+    }}
+  >
     <div className="px-8 py-4">
       <div className="mt-8 text-xl font-medium text-gray-800 leading-1 sm:mt-0 sm:pr-32 font-heading">
         {title}
@@ -33,7 +41,7 @@ export const ConfirmationPrompt: FC<ConfirmationPromptProps> = ({
           'mt-8 sm:-mt-2': !!icon && !!description,
         })}
       >
-        <Button theme="secondary" size="base" className="shrink-0 mr-5" onClick={onRefuse}>
+        <Button theme="secondary" size="base" className="mr-5 shrink-0" onClick={onRefuse}>
           No
         </Button>
         <Button theme="primary" size="base" className="shrink-0 sm:mr-5" onClick={onAccept}>
