+++ src/components/modal/content/types.ts
@@ -1,6 +1,9 @@
-import { ModalProps } from '..';
+import { ModalProps } from '../types';
 
 export type ModalContentProps = ModalProps & {
   /** Size of the viewport in which the modal is rendered */
   viewport: string;
+
+  floating: any;
+  getFloatingProps: any;
 };
