+++ src/components/modal/types.d.ts
@@ -17,19 +17,15 @@ export interface ModalProps {
   /**
    * Size (width) of the modal
    */
-  size?: 'narrow' | 'default' | 'wide';
+  size?: 'xs' | 's' | 'default' | 'l' | 'xl';
+
   children?: ReactNode;
   /**
    * Class name to assign to the modal
    */
   className?: string;
   /**
-   * Whether the modal's content should be scrollable if it overflows. Default to `true`.
+   * Callback executed when the modal open change
    */
-  scrollable?: boolean;
-  /**
-   * Callback executed when the modal is dismissed by clicking on the overlay, the close button or
-   * pressing the escape key
-   */
-  onDismiss: () => void;
+  onOpenChange: (open: boolean) => void;
 }
