--- src/components/modal2/content/index.ts
@@ -1,2 +1,0 @@
-export type { ModalContentProps } from './types';
-export { default } from './component';
