--- src/components/modal2/index.ts
@@ -1,2 +1,0 @@
-export { default } from './component';
-export type { ModalProps } from './types';
