--- src/components/tooltip/arrow/types.d.ts
@@ -1,3 +1,0 @@
-export interface ArrowProps {
-  'data-placement'?: string;
-}
